made enemy dive cluster types

+++ levelLoader.py
@@ -143,14 +143,13 @@ class LevelLoader():
         '''contructs and returns enemies based off a 1 input nameing convention'''
         ''' can add enemy classes, just contruct new enemy type'''
         enemySprite = None
-
         if className[0]=="@": #quick naming scheme ex: "@s1-d-3-1" -check levelLibrary.py for more
             a=className[1::].split("-")
             if PROPER_FORMAT(className):
                 if len(a)==4:
                     enemySprite = enemy.enemy(ENEMY_SECTORS(a[0]), behavior=ENEMY_TYPE_MAP(a[1]), speed=int(a[2]), health=int(a[3]))
-                else:
-                    enemySprite = enemy.enemy(ENEMY_SECTORS(a[0]), behavior=ENEMY_TYPE_MAP(a[1]), speed=int(a[2]), health=int(a[3]), acceleration=int(a[4]))
+                else:     
+                    enemySprite = enemy.enemy(ENEMY_SECTORS(a[0]), behavior=ENEMY_TYPE_MAP(a[1]), speed=int(a[2]), health=int(a[3]), acceleration=float(a[4]))
                 return enemySprite
